pyproject.toml
[project]
name = "trecl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "langchain-cerebras>=0.8.2",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langsmith>=0.7.9",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "tavily-python>=0.7.22",
    "tenacity>=9.1.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "health: tests that require live API connections (deselect with '-m not health')",
]
